Separate test for local and cloud blob transfer. Move common code to abstract class
