+++ system-tests/src/test/java/org/eclipse/dataspaceconnector/system/tests/local/AbstractBlobTransferTest.java
@@ -1,0 +1,58 @@
+package org.eclipse.dataspaceconnector.system.tests.local;
+
+import com.azure.storage.blob.BlobServiceClient;
+import com.azure.storage.blob.BlobServiceClientBuilder;
+import com.azure.storage.common.StorageSharedKeyCredential;
+import io.restassured.response.ResponseBodyExtractionOptions;
+import org.eclipse.dataspaceconnector.system.tests.utils.TransferSimulationUtils;
+import org.jetbrains.annotations.NotNull;
+
+import static io.restassured.RestAssured.given;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.eclipse.dataspaceconnector.system.tests.local.BlobTransferLocalSimulation.ACCOUNT_NAME_PROPERTY;
+import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.API_KEY;
+import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.API_KEY_HEADER;
+import static org.eclipse.dataspaceconnector.system.tests.local.TransferLocalSimulation.CONSUMER_MANAGEMENT_URL;
+import static org.eclipse.dataspaceconnector.system.tests.utils.GatlingUtils.runGatling;
+import static org.eclipse.dataspaceconnector.system.tests.utils.TransferSimulationUtils.PROVIDER_ASSET_FILE;
+import static org.eclipse.dataspaceconnector.system.tests.utils.TransferSimulationUtils.TRANSFER_PROCESSES_PATH;
+
+public abstract class AbstractBlobTransferTest {
+
+
+    protected void initiateTransfer(BlobServiceClient dstBlobServiceClient) {
+        // Act
+        System.setProperty(ACCOUNT_NAME_PROPERTY, dstBlobServiceClient.getAccountName());
+        runGatling(BlobTransferLocalSimulation.class, TransferSimulationUtils.DESCRIPTION);
+
+        // Assert
+        var container = getProvisionedContainerName();
+        var destinationBlob = dstBlobServiceClient.getBlobContainerClient(container)
+                .getBlobClient(PROVIDER_ASSET_FILE);
+        assertThat(destinationBlob.exists())
+                .withFailMessage("Destination blob %s not created", destinationBlob.getBlobUrl())
+                .isTrue();
+    }
+
+    @NotNull
+    protected BlobServiceClient getBlobServiceClient(String endpoint, String accountName, String accountKey) {
+
+        return new BlobServiceClientBuilder()
+                .endpoint(endpoint)
+                .credential(new StorageSharedKeyCredential(accountName, accountKey))
+                .buildClient();
+    }
+
+    private String getProvisionedContainerName() {
+        ResponseBodyExtractionOptions body = given()
+                .baseUri(CONSUMER_MANAGEMENT_URL)
+                .header(API_KEY_HEADER, API_KEY)
+                .when()
+                .get(TRANSFER_PROCESSES_PATH)
+                .then()
+                .statusCode(200)
+                .extract().body();
+        return body
+                .jsonPath().getString("[0].dataDestination.properties.container");
+    }
+}
